+++ crypto/src/aes/aes_gcm_arm64.rs
@@ -43,7 +43,7 @@ pub(crate) unsafe fn gcm_encrypt_armv8<const N: usize>(
     }
 
     assert!(
-        in_out.len() <= GCM_MAX_LEN,
+        in_out.len() as u64 <= GCM_MAX_LEN,
         "GCM plaintext exceeds maximum allowed length (2^32 - 2 blocks)"
     );
